sfio sfprint: squash -Wsign-compare warning in 'lv' comparison
This comparison is trying to check the high bit of `lv` which can be done in
`unsigned long long` space.

+++ lib/sfio/sfprint.c
@@ -465,7 +465,7 @@ int sfprint(FILE *f, Sffmt_t *format) {
 		    break;
 		if (lv < 0 && fmt == 'd') {
 		    flags |= SFFMT_MINUS;
-		    if (lv == HIGHBITL) {	/* avoid overflow */
+		    if ((unsigned long long)lv == HIGHBITL) { // avoid overflow
 			lv = (long long)(HIGHBITL / base);
 			*--sp = _Sfdigits[HIGHBITL - (unsigned long long)lv * base];
 		    } else
